backend/requirements.txt
# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# ML / pose
ultralytics==8.2.87
opencv-python-headless==4.10.0.84
numpy==1.26.4

# ML training & inference
scikit-learn==1.5.2
joblib==1.4.2

# Supabase & auth
supabase==2.7.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Config
pydantic-settings==2.4.0
pydantic[email]>=2.0
python-dotenv==1.0.1

# HTTP (used for Supabase Storage uploads/downloads)
requests>=2.31.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
